Первая версия ДЗ 7-го модуля - попытка вторая
